pyproject.toml
[tool.poetry]
name = "Kwazii_doc_parser"
version = "0.0.1"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
ipykernel = "^6.29.3"
python-dotenv = "1.0.1"
langchain = "0.3.13"
langgraph = "^0.2.70"
langchain-openai = "0.2.14"
langchain-ollama = "0.2.2"
langchain-anthropic = "0.3.1"
langchain-upstage = "0.4.0"
langchain-cohere = "0.3.4"
langchain-milvus = "0.1.7"
langchain-google-genai = "2.0.7"
langchain-huggingface = "0.1.2"
langchain-together = "0.2.0"
faiss-cpu = "1.9.0.post1"
ragas = "0.2.9"
jupyter = "1.1.1"
notebook = "7.3.2"
chardet = "5.2.0"
markitdown = "0.0.1a3"
pypdf = "4.3.1"
llama-parse = "0.5.19"
arxiv = "^2.1.3"
pymupdf = "^1.25.1"
neo4j = "^5.27.0"
langchain-neo4j = "^0.2.0"
langchain-mongodb = "^0.3.0"
certifi = "^2024.12.14"
python-magic-bin = { version = "^0.4.14", markers = "sys_platform != 'darwin'" }
pymongo = "^4.10.1"
langchain-qdrant = "^0.2.0"
chromadb = "^0.5.0"
pyalex = "^0.15.1"
pdfplumber = "^0.11.5"
unstructured = {extras = ["all-docs"], version = "^0.16.13"}
openai = "^1.59.7"
langchain-experimental = "^0.3.4"
tiktoken = "^0.8.0"
pytesseract = "^0.3.13"
lxml = "^5.3.0"
pillow = "^11.1.0"
pydantic = "^2.10.5"
matplotlib = "^3.10.0"
feedparser = "^6.0.11"
tavily-python = "^0.5.0"
pandas = "^2.2.3"
pinecone = {version = "5.0.1", extras = ["grpc"]}
langchain-elasticsearch = "^0.3.1"
anytree = "2.12.1"
langchain-chroma = "^0.2.0"
olefile = "^0.47"
pdf2image = "^1.17.0"
beautifulsoup4 = "^4.12.3"
mlx = {version = "^0.22.0", markers = "sys_platform == 'darwin'"}
mlx-lm = {version = "^0.21.1", markers = "sys_platform == 'darwin'"}
seaborn = "^0.13.2"
langchain-unstructured = "^0.1.6"
rapidocr-onnxruntime = "^1.4.4"
ollama = "^0.4.8"
groq = "^0.22.0"
together = "^1.5.5"
cerebras-cloud-sdk = "^1.29.0"
anthropic = "^0.49.0"
google-generativeai = "^0.8.5"
docling = "^2.30.0"
pymupdf4llm = "^0.0.21"
rank-bm25 = "^0.2.2"
nest-asyncio = "^1.6.0"
llama-cloud-services = "^0.6.51"
requests = "^2.32.4"
voyageai = "^0.2.0"
qdrant-client = "^1.7.0"
fastembed = "^0.7.1"
numpy = "^1.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

